[0.0.35] Added script validation error for deleting allowed process creations

+++ src/index.ts
@@ -169,6 +169,11 @@ try {
           cwd,
         });
 
+        const validateScriptError = validateScript(script);
+        if (validateScriptError) {
+          return validateScriptError;
+        }
+
         await deleteAllowedProcessCreation({
           script,
           args,
